requirements.txt
flet
opencv-python
pyodbc
opencv-contrib-python==4.9.0.80
openpyxl
pillow
pandas